Clean up code

+++ src/main/java/org/assimbly/gateway/web/rest/broker/HealthBrokerResource.java
@@ -88,7 +88,7 @@ public class HealthBrokerResource {
             backendResponse.addMemory("current", healthService.convertSizeToKb(mem.getUsed()));
             backendResponse.addMemory("max", healthService.convertSizeToKb(mem.getMax()));
             backendResponse.addMemory("committed", healthService.convertSizeToKb(mem.getCommitted()));
-            backendResponse.addMemory("cached", healthService.convertSizeToKb(mem.getCommitted() - mem.getUsed()));
+            backendResponse.addMemory("cached", healthService.convertSizeToKb(mem.getCommitted() - (double) mem.getUsed()));
             backendResponse.addMemory("currentUsedPercentage", (mem.getUsed() * 100 / mem.getMax()));
 
             backendResponse.addThread("threadCount", threadMXBean.getThreadCount());
